Added Upsert and UpsertTTL to support atomic updates backed by WATCH and Transactions
Adds Upsert and UpsertTTL functions that allow for a key to be retrieved and updated within a transaction. If the key is changed during a transaction Redis will fail the transaction (optimistic locking model). A new RetryableError type and IsRetryable function has been added so that the caller can determine if the error returned from Upsert is retryable. Anytime the transaction fails from Redis end a RetryableError is returned so the caller can retry.