Additional metrics and bug fixes
Adds a new prometheus metric to track how many keys are being requested for MGET operations. This can be useful to understand how much pressure may be put against the Redis proxy when running in Enterprise Clustering, or how much pressure will be put on the nodes when not running in OSS clustering policy.

Fixes a bug where Prometheus cache hits metric was being incremented when there was a cache miss.

Moves the version to 1.5.1

Signed-off-by: Joseph Kratz <4985721+jkratz55@users.noreply.github.com>